+++ backend/modules/context_size_calculator_lib/tokenizers/microsoft_Phi-4-mini-instruct/chat_template.jinja
