+++ backend/modules/context_size_calculator_lib/tokenizers/mistralai_Mistral-7B-Instruct-v0.3/chat_template.jinja
